modify P1

+++ test/test_project3/add.c
@@ -22,7 +22,7 @@ int main(int argc, char * argv[])
     assert(argc >= 2);
 
     
-    struct TestMultiCoreArg *args = (struct TestMultiCoreArg *)(atoi(argv[1]));
+    struct TestMultiCoreArg *args = (struct TestMultiCoreArg *)(atol(argv[1]));
 
     int print_location = args->print_location;
     int from  = args->from;
