+++ test/test_project3/producer.c
@@ -16,7 +16,7 @@ int main(int argc, char *argv[])
     int print_location = atoi(argv[1]);
     int handle_cond = atoi(argv[2]);
     int handle_lock = atoi(argv[3]);
-    int * num_staff = (int *)(atoi(argv[4])); 
+    int * num_staff = (int *)(atol(argv[4])); 
 
     // Set random seed
     srand(clock());
